feat: Add APIQualityMiddleware for handling API quality
