+++ Sources/App/Controllers/APIController.swift
@@ -13,38 +13,9 @@ enum APIServiceError: Error {
     case unknown
 }
 
-/// API の品質を表現します。
-enum APIQuality {
-    case sometimesFails(probability: Double)
-    case alwaysFails
-    case neverFails
-
-    func test() -> Bool {
-        return switch self {
-        case .sometimesFails(let probability):
-            shouldFail(probability: probability)
-        case .alwaysFails:
-            true
-        case .neverFails:
-            false
-        }
-    }
-
-    private func shouldFail(probability: Double) -> Bool {
-        return (0 ..< probability).contains(Double.random(in: 0 ..< 1))
-    }
-}
-
 struct APIController: APIProtocol {
 
-    /// API の想定品質
-    var apiQuality: APIQuality?
-
     func post_hyphen_my_hyphen_fortune(_ input: Operations.post_hyphen_my_hyphen_fortune.Input) async throws -> Operations.post_hyphen_my_hyphen_fortune.Output {
-
-        if let apiQuality, apiQuality.test() {
-            throw APIServiceError.unknown
-        }
 
         let version = input.headers.API_hyphen_Version ?? .v1
         switch version {
