+++ Sources/App/configure.swift
@@ -7,8 +7,11 @@ public func configure(_ app: Application) async throws {
     app.middleware.use(
         FileMiddleware(publicDirectory: app.directory.publicDirectory, defaultFile: "index.html")
     )
+    app.middleware.use(
+        APIQualityMiddleware(apiQuality: .sometimesFails(probability: 0.25), onError: APIServiceError.unknown)
+    )
     // register routes
     let transport = VaporTransport(routesBuilder: app)
-    let controller = APIController(apiQuality: .sometimesFails(probability: 0.25))
+    let controller = APIController()
     try controller.registerHandlers(on: transport)
 }
